backend/requirements.txt
flask
flask-cors
transformers
pillow
torch
markitdown[all]